+++ cf/test/cfa_test.sh
@@ -1,8 +1,7 @@
 # --------------------------------------------------------------------
 # Test cfa
 # --------------------------------------------------------------------
-scripts=../../scripts
-sample_files=../../docs/source/sample_files
+sample_files=$PWD
 
 test_file=delme_cfa.nc
 test_dir=delme_cfa_dir
@@ -11,22 +10,28 @@ mkdir -p $test_dir
 
 for opt in vs vm vc
 do
-#  echo $opt
-  $scripts/cfa    -$opt $sample_files/*[np][cp] >/dev/null
-  $scripts/cfa -1 -$opt $sample_files/*[np][cp] >/dev/null
-  for f in $sample_files/*[np][cp]
+  cfa    -$opt $sample_files/*[np][cp] >/dev/null
+  cfa -1 -$opt $sample_files/*[np][cp] >/dev/null
+  for f in `ls $sample_files/*.[np][cp] | grep -v $test_file`
   do
 #    echo $f
-    $scripts/cfa -$opt $f >/dev/null
+    cfa -$opt $f >/dev/null
     if [ $opt = vs ] ; then
-      $scripts/cfa --overwrite -o $test_file $f
-      $scripts/cfa --overwrite -d $test_dir  $f
+      cfa --overwrite -o $test_file $f
+      cfa --overwrite -d $test_dir  $f
     fi
   done
+  rm -f $test_file
 done
 
-$scripts/cfa --overwrite -o $test_file $sample_files/*[np][cp]
-$scripts/cfa --overwrite -d $test_dir  $sample_files/*[np][cp]
+rm -f $test_file
+cfa --overwrite -d $test_dir  $sample_files/*.[np][cp]
+
+rm -f $test_file
+cfa -o $test_file $sample_files/test*.[np][cp]
+
+rm -f $test_file
+cfa -o $test_file $test_dir
 
 rm -fr $test_dir $test_file